Address review feedback: fix double margins, deduplicate PDF utils, include user info in mobile PDF
- Remove body margin (keep @page margin only) to prevent double margins in expo-print
- Extract shared PDF constants and helpers into pdfUtils.ts
- Always include userInfo in mobile PDF to match web behavior
